+++ src/constants.mjs
@@ -7,3 +7,4 @@ export const PROJECT_ROOT = path.join(__dirname, '..');
 export const SCRIPT_HELPER = path.join(PROJECT_ROOT, 'src', 'script-helper.js');
 export const EXTENSION_PATH = path.join(PROJECT_ROOT, 'build');
 export const CACHE_DIR_NAME = '.jambox';
+export const CONFIG_FILE_NAME = 'jambox.config.js';